Fix bug with clear_lines
ChatGPT prompt that fixed this:
One of these two functions has a bug that upon line clears is occasionally causing the entire vertical playfield above the played piece to be filled with the piece's color. Which of these functions is causing this bug, why, and how should the function be rewritten in PICO8?